Fix: Runtime error for AppStoreReview
This is because RNAppStoreReview is not getting added to the build -
it's not appearing in the Library search paths and ldflags via
Pods/*.xcconfig

Looking into this it's because the podspec file in not in the root
directory and the sources property is referencing a path that does not
exist. Too bad pod install doesn't alert us to this problem!

Moved the podspec to root.

changed:
s.source_files  = "RNAppStoreReview/**/*.{h,m}"
To:
s.source_files  = "ios/**/*.{h,m}"

Updated readme

modified:   README.md
renamed:    ios/RNAppStoreReview.podspec -> RNAppStoreReview.podspec